[IMP] rpgle-parser: Add AST layer
The CST was unable to provide sufficient context to even fully
tokenize in the manner I needed. Specifically, I want to have
literals be a single token, but in rpgle they can span multiple
lines, but only in specific fields due to the fixed width. This
commit introduces an AST that loses a bit of context (comments,
blanklines), but can be used to incorporate continuation lines.
This will be the preferred layer for free parsing going forward.