[BUILD]: add dependency on nonempty crate
Some parsing activites are causing a relative loss of semantic context.
One specific example is conversion of a known width character array into
a character vector. The vector is known to have values, so this commit
upgrades these vectors to NonEmpty<T> so we can guarantee at least
to have one row/column to pass upward.