[IMP] idk-nvim,rpgle-parser: first search current,then search external
When searching an external program name using goto definition, the
search was getting stuck at the prototype definition in the current file
and would not search through the other rpg files. This commit adds code
to identify if you are in a prototype definition and if so, will search
the external files and open that file.